week02/day10/requirements.txt
faiss-cpu
sentence-transformers
rank-bm25
numpy
pandas
scikit-learn
tqdm
streamlit
matplotlib
